+++ packages/server/src/_utils/models/root/api_key_permission.ts
@@ -1,4 +1,0 @@
-export class ApiKeyPermission {
-  api_key_id: string;
-  permission_id: number | bigint;
-}