+++ packages/server/src/_utils/models/root/database.ts
@@ -1,5 +1,7 @@
-export class Database {
-  id: number | bigint;
+import { ChiselId } from '@capsulesh/shared-types';
+
+export class DatabaseModel {
+  id: ChiselId;
   name: string;
   client_id: string;
 }