+++ packages/server/src/_utils/models/root/oauth_clients.ts
@@ -7,3 +7,10 @@ export class OauthClients {
   created_at: Date;
   updated_at: Date | null;
 }
+
+export class ClientModel {
+  id: number;
+  name: string;
+  redirect_uri: string;
+  organization_id?: number;
+}