+++ packages/server/src/persons/dto/response/get-person.dto.ts
@@ -1,13 +1,9 @@
-import { GetPerson } from '@capsulesh/shared-types';
-
-export class GetPersonDto implements GetPerson {
+export class GetPersonDto {
   id: number;
   firstName: string;
   lastName: string;
   dateOfBirth: string;
-  gender: string;
   email: string;
   phone: string;
-  address: string;
   occupation: string;
 }