+++ packages/server/src/tasks/tasks.service.ts
@@ -1,7 +1,9 @@
 import { Injectable } from '@nestjs/common';
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import { TaskEvents } from 'src/_utils/decorators/event-emitter.decorator';
+import { ResourceTypeEnum } from 'src/_utils/models/root/resource';
 import { QueryOptionsDto } from 'src/dynamic-queries/_utils/dto/request/query-options.dto';
+import { ResourcesRepository } from 'src/resources/resources.repository';
 import { CreateTaskDto } from './dto/request/create-task.dto';
 import { UpdateTaskDto } from './dto/request/update-task.dto';
 import { TasksMapper } from './tasks.mapper';
@@ -13,13 +15,27 @@ export class TasksService {
     private readonly eventEmitter: EventEmitter2,
     private readonly tasksRepository: TasksRepository,
     private readonly tasksMapper: TasksMapper,
+    private readonly resourcesRepository: ResourcesRepository,
   ) {}
+  private readonly RESOURCE_TASK_PREFIX = 'task';
 
   @TaskEvents.emitCreated()
   async createTask(createTaskDto: CreateTaskDto) {
-    const result = this.tasksRepository.createTask(createTaskDto);
-    const createdTask = this.tasksRepository.findTaskById(result.id);
-    return this.tasksMapper.toGetTaskDto(createdTask);
+    const resource = this.resourcesRepository.createResource(
+      this.RESOURCE_TASK_PREFIX + Date.now().toString(),
+      ResourceTypeEnum.TASK,
+    );
+
+    if (resource.id) {
+      const result = this.tasksRepository.createTask(
+        createTaskDto,
+        resource.id as number,
+      );
+      console.log('Result', resource, result);
+      const createdTask = this.tasksRepository.findTaskById(resource.id);
+      console.log('Created Task', createdTask);
+      return this.tasksMapper.toGetTaskDto(createdTask);
+    }
   }
 
   @TaskEvents.emitUpdated()